now the sidebar handles the displayed papers
